move MarkOldConventionAsDeprecated to usecase folder
